يا علي مدد/requirements.txt
# Core Dependencies
requests==2.31.0
urllib3==2.1.0
aiohttp==3.9.0
httpx==0.25.0

# Async Support
asyncio==3.4.3
gevent==23.9.1

# HTML/XML Parsing
beautifulsoup4==4.12.2
lxml==4.9.3
html5lib==1.1

# Network & DNS
dnspython==2.4.2
scapy==2.5.0
python-nmap==0.7.1

# Database
SQLAlchemy==2.0.23

# AI/ML Libraries
openai==1.3.0
anthropic==0.7.0
transformers==4.35.0
torch==2.1.0
langchain==0.0.335

# Encoding & Cryptography
cryptography==41.0.7
pycryptodome==3.19.0
base58==2.1.1

# JWT Handling
PyJWT==2.8.0
python-jose==3.3.0

# YAML & JSON
PyYAML==6.0.1
jsonschema==4.20.0

# Report Generation
jinja2==3.1.2
markdown==3.5.1
reportlab==4.0.7
weasyprint==60.1
matplotlib==3.8.2
plotly==5.18.0

# Color & Output
colorama==0.4.6
termcolor==2.3.0
rich==13.7.0
tqdm==4.66.1

# Proxy & Session
requests-toolbelt==1.0.0
PySocks==1.7.1

# Configuration
python-dotenv==1.0.0
configparser==6.0.0

# Date & Time
python-dateutil==2.8.2
pytz==2023.3

# URL Parsing
validators==0.22.0
furl==2.1.3

# Logging
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Cloud APIs
boto3==1.29.7
azure-identity==1.15.0
google-cloud-storage==2.10.0

# Notifications
slack-sdk==3.23.0
discord-webhook==1.3.0
python-telegram-bot==20.7

# Image Processing
Pillow==10.1.0

# Regular Expressions
regex==2023.10.3

# Rate Limiting
ratelimit==2.2.1

# Random Data Generation
Faker==20.1.0

# WAF Detection
wafw00f==2.2.0

# Subdomain Enumeration
sublist3r==1.1

# Hash Handling
passlib==1.7.4

# GraphQL
gql==3.4.1
graphql-core==3.2.3

# WebSockets
websockets==12.0

# File Type Detection
python-magic==0.4.27

# PDF Processing
PyPDF2==3.0.1

# Excel Processing
openpyxl==3.1.2

# Email
email-validator==2.1.0

# User Agents
fake-useragent==1.4.0

# Selenium (for DOM XSS)
selenium==4.15.0
webdriver-manager==4.0.1

# System Information
psutil==5.9.6

# Progress Bars
alive-progress==3.1.5

# QR Code (for 2FA Testing)
qrcode==7.4.2

# TOTP (for 2FA Testing)
pyotp==2.9.0

# Image Comparison
imagehash==4.3.1

# Machine Learning
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3

# Natural Language Processing
nltk==3.8.1

# Version Control
GitPython==3.1.40

# Archive Handling
py7zr==0.20.8

# Code Quality
black==23.11.0
pylint==3.0.2
flake8==6.1.0

# Performance Monitoring
memory-profiler==0.61.0